Added server and client side validation of edit and remove payment method

+++ App/Models/ExpenseModel.php
@@ -358,28 +358,34 @@ class ExpenseModel extends \Core\Model {
     }
 
     public static function deletePaymentAssignedToUser($method) {
-        $sql = 'DELETE FROM payment_methods_assigned_to_users
-                WHERE user_id = :user_id AND name = :method';
+        if (static::findPaymentMethodAssignedToUser($method)) {
+            $sql = 'DELETE FROM payment_methods_assigned_to_users
+                    WHERE user_id = :user_id AND name = :method';
 
-        $db = static::getDB();
-        $stmt = $db->prepare($sql);
-        $stmt->bindValue(':user_id', $_SESSION['user_id'], PDO::PARAM_INT);
-        $stmt->bindValue(':method', $method, PDO::PARAM_STR);
+            $db = static::getDB();
+            $stmt = $db->prepare($sql);
+            $stmt->bindValue(':user_id', $_SESSION['user_id'], PDO::PARAM_INT);
+            $stmt->bindValue(':method', $method, PDO::PARAM_STR);
 
-        return $stmt->execute();
+            return $stmt->execute();
+        }
+        return false;
     }
 
     public static function updatePaymentAssignedToUser($method, $newMethod) {
-        $sql = 'UPDATE payment_methods_assigned_to_users
-                SET name = :new_method
-                WHERE name = :method AND user_id = :user_id';
+        if (static::findPaymentMethodAssignedToUser($method)) {
+            $sql = 'UPDATE payment_methods_assigned_to_users
+                    SET name = :new_method
+                    WHERE name = :method AND user_id = :user_id';
 
-        $db = static::getDB();
-        $stmt = $db->prepare($sql);
-        $stmt->bindValue(':new_method', $newMethod, PDO::PARAM_STR);
-        $stmt->bindValue(':method', $method, PDO::PARAM_STR);
-        $stmt->bindValue(':user_id', $_SESSION['user_id'], PDO::PARAM_INT);
+            $db = static::getDB();
+            $stmt = $db->prepare($sql);
+            $stmt->bindValue(':new_method', $newMethod, PDO::PARAM_STR);
+            $stmt->bindValue(':method', $method, PDO::PARAM_STR);
+            $stmt->bindValue(':user_id', $_SESSION['user_id'], PDO::PARAM_INT);
 
-        return $stmt->execute();
+            return $stmt->execute();
+        }
+        return false;
     }
 }
